register and sign in refactoring
